+++ package.json
@@ -20,7 +20,6 @@
     "@radix-ui/react-radio-group": "^1.2.0",
     "@radix-ui/react-select": "^2.1.1",
     "@radix-ui/react-slot": "^1.1.0",
-    "@sendgrid/mail": "^8.1.3",
     "@tanstack/react-table": "^8.20.5",
     "@react-google-maps/api": "^2.19.3",
     "@sendgrid/mail": "^8.1.3",
@@ -43,7 +42,6 @@
     "react-dropzone": "^14.2.3",
     "react-hook-form": "^7.53.0",
     "react-icons": "^5.3.0",
-    "react-phone-input-2": "^2.15.1"
     "react-phone-input-2": "^2.15.1",
     "react-phone-number-input": "^3.4.7",
     "react-router-dom": "^6.26.2",
